Big refactor
