--- twine-components/src/example.rs
@@ -1,3 +1,0 @@
-pub mod area_calculators;
-pub mod math;
-pub mod oscillator;
