--- twine-components/src/example/math.rs
@@ -1,9 +1,0 @@
-mod adder;
-mod arithmetic;
-mod min_max;
-mod multiplier;
-
-pub use adder::Adder;
-pub use arithmetic::{Arithmetic, ArithmeticInput, ArithmeticOutput};
-pub use min_max::{MinMax, MinMaxInput, MinMaxOutput};
-pub use multiplier::Multiplier;
