--- twine-components/src/fluid.rs
@@ -1,5 +1,0 @@
-mod ideal_gas;
-mod incompressible_liquid;
-
-pub use ideal_gas::{IdealGasModel, IdealGasState};
-pub use incompressible_liquid::IncompressibleLiquid;
