+++ twine-components/src/lib.rs
@@ -1,7 +1,3 @@
 pub mod controller;
-pub mod example;
-pub mod fluid;
-pub mod integrators;
-pub mod interpolation;
 pub mod schedule;
 pub mod thermal;
