+++ twine-components/src/thermal.rs
@@ -1,2 +1,1 @@
 pub mod tank;
-pub mod tank2;
