+++ twine-core/src/lib.rs
@@ -1,12 +1,7 @@
-mod component;
 pub mod constraint;
 pub mod graph;
 mod simulation;
-pub mod thermo;
 mod time;
-mod twine;
 
-pub use component::Component;
-pub use simulation::{Simulation, State};
+pub use simulation::{Model, Simulation, State};
 pub use time::{DurationExt, TimeDerivative, TimeIntegrable};
-pub use twine::{Twine, TwineError};
