--- twine-core/src/thermo.rs
@@ -1,2 +1,0 @@
-pub mod fluid;
-pub mod units;
